requirements.txt
fastapi
uvicorn
playwright
python-dotenv
pydantic
httpx
pytest
